requirements.txt
GitPython
schedule
python-dotenv